third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-import-block-but-nonce-nested-expected.txt

CONSOLE ERROR: Refused to load the script 'http://localhost:8080/htmlimports/resources/external-script.js' because it violates the following Content Security Policy directive: "script-src http://localhost:8000 'nonce-hello'".

CONSOLE ERROR: line 4: Uncaught EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src http://localhost:8000 'nonce-hello'".

CONSOLE ERROR: line 6: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://localhost:8000 'nonce-hello'". Either the 'unsafe-inline' keyword, a hash ('sha256-wnLyUqsXhUuQ5NVNJO+JjNabqB1yo9OPLLJWPopauN0='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 13: Uncaught EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src http://localhost:8000 'nonce-hello'".

CONSOLE MESSAGE: line 12: PASS
CONSOLE MESSAGE: line 12: PASS
CONSOLE MESSAGE: line 12: PASS
CONSOLE MESSAGE: line 12: PASS
CONSOLE MESSAGE: line 12: PASS
CONSOLE MESSAGE: line 12: PASS
CONSOLE MESSAGE: line 12: PASS
CONSOLE MESSAGE: line 12: PASS
